requirements.txt
flask
flask_session
werkzeug.security
requests
sqlite3
re